deleted component functions directory and included function in compressor component
